renamed addtoBag to addtoCart
