+++ src/components/ProductDetail/StylesAndSizes/SizesAndAddToCart/CartAndOutfit/AddToCart.jsx
@@ -2,11 +2,11 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import styles from '../styles.css';
 
-function AddToBag({
+function AddToCart({
   currentSize, currentQuantity, productID, styleID, purchasePrice,
 }) {
   return (
-    <div data-testid="addToBag" id={styles.addToBag}>
+    <div data-testid="addToCart" id={styles.addToBag}>
       <button
         type="button"
         id={styles.addToBagButton}
@@ -16,15 +16,15 @@ function AddToBag({
           });
         }}
       >
-        Add To Bag
+        Add To Cart
       </button>
     </div>
   );
 }
 
-export default AddToBag;
+export default AddToCart;
 
-AddToBag.propTypes = {
+AddToCart.propTypes = {
   currentSize: PropTypes.string.isRequired,
   currentQuantity: PropTypes.number.isRequired,
   productID: PropTypes.string.isRequired,
